Initial Authorization Levels (#7)
* Updated migrations, seeders and factories to support user types

* Added a policy so that only admins can add and delete teams

* Added gates to differentiate admin and standard user views

* Teams delete function now utilises type hinting

+++ app/Http/Controllers/TeamController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use App\Models\User;
 use App\Models\Team;
 use App\Models\City;
 use App\Models\Venue;
@@ -31,8 +32,14 @@ class TeamController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request)
+    public function store(Request $request, Team $team)
     {
+
+        # Only allow admins to add teams
+        if ($request->user()->cannot('create', $team)) {
+            abort(403);
+        }
+
         //Needs to validate that city and venue ids exist
         $validatedData = $request->validate([
             'name' => 'required|max:255',
@@ -77,9 +84,13 @@ class TeamController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id)
+    public function destroy(Request $request, Team $team)
     {
-        $team = Team::findOrFail($id);
+        # Only allow admins to delete teams
+        if ($request->user()->cannot('delete', $team)) {
+            abort(403);
+        }
+
         $team->delete();
 
         # Flash message
